+++ modules/se_sdk3_hosting/ProcessorStateManager.cpp
@@ -196,7 +196,7 @@ void DawPreset::initFromXML(const std::map<int32_t, paramInfo>& parametersInfo, 
 				formatVersion = 1;
 				try
 				{
-					fourCC = std::stoul(hexcode, nullptr, 16);
+					fourCC = static_cast<int32_t>(std::stoul(hexcode, nullptr, 16));
 				}
 				catch (...)
 				{
